feat(gate6a): pull-only — delete wake dispatcher/adapters/reachability; senders stop poking
Gate 6a of the Hybrid migration — "senders write files; they never wake recipients."
-2624/+187 LOC, 12 files deleted. Tree green, -race clean, conformance green. run.go
touched only at the reprove-call removal (1 hunk, 9 deletions); the PTY/submit-bytes/
termios/SIGWINCH code + the 3 exact-byte submit tests are untouched.

DELETE: internal/loop/wake.go (WakeAdapter/registry/PokeRegistration/PokeWakeTarget +
Set*Hook seams + tmuxAdapter), reachability.go (reproveAndRebindOwnWake + herdr/tmux
reprove), internal/loop/herdr.go + tmux.go (poke adapters), runner_reachable.go (hook
injection) + all their tests.
CO-TRIM (intended pull-only removal): send.go/defer.go/message.go drop PokeRegistration
(deliver by file write only; send wake-receipt -> constant "none (pull)"); doctor drops
checkWakeTargetValidity + checkRunnerSocketStaleness (probed via deleted hooks);
run.go:685 drop the reprove call; poller.go:220 drop its reprove call.
RELOCATE (verbatim, to avoid editing Gate-6c-reserved register.go): reproveProbeTimeout
-> runner_provenance.go; registrationHasReachableWake -> poller.go; test helpers ->
test_helpers_test.go. RegistrationReachable kept but reimplemented runner-receive-socket-
only (status/register call it; the receive socket survives until 6b).

VESTIGIAL, removed later: run.go receive-socket (6b); wake_method/wake_target fields +
register wake-autodetect/tmux-pane-dedup + tmux_state.go/herdr_state.go (6c).

Compile gate: no PokeRegistration/PokeWakeTarget/WakeAdapter/reproveAndRebindOwnWake/
SetHerdr/SetTmux refs remain.
